requirements.dev.txt
django-extra-checks
django-querycount
django-silk
flake8
flake8-bandit
flake8-django
flake8-pytest
flake8-pytest-style
freezegun
model_bakery
nplusone
pytest
pytest-cov
pytest-deadfixtures
pytest-django
pytest-randomly
pytest-testmon
pytest-timeout
